# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Romeo PHD 6.0 - Agent Development Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type PipelineStatus =
  (typeof PipelineStatus)[keyof typeof PipelineStatus];

export const PipelineStatus = {
  pending: "pending",
  running: "running",
  paused: "paused",
  completed: "completed",
  failed: "failed",
} as const;

export interface Pipeline {
  id: number;
  name: string;
  yamlContent: string;
  status: PipelineStatus;
  nodeCount: number;
  resolvedCount: number;
  createdAt: string;
  updatedAt: string;
}

export type PipelineNodeStatus =
  (typeof PipelineNodeStatus)[keyof typeof PipelineNodeStatus];

export const PipelineNodeStatus = {
  PENDING: "PENDING",
  RESOLVED: "RESOLVED",
  NEEDS_CLARIFICATION: "NEEDS_CLARIFICATION",
  AMBIGUOUS: "AMBIGUOUS",
} as const;

export interface PipelineNode {
  id: number;
  pipelineId: number;
  nodeId: string;
  name: string;
  type: string;
  /** @nullable */
  prompt?: string | null;
  dependencies: string[];
  status: PipelineNodeStatus;
  /** @nullable */
  output?: string | null;
  /** @nullable */
  confidenceScore?: number | null;
  positionX: number;
  positionY: number;
  /** @nullable */
  executedAt?: string | null;
}

export type PipelineWithNodesStatus =
  (typeof PipelineWithNodesStatus)[keyof typeof PipelineWithNodesStatus];

export const PipelineWithNodesStatus = {
  pending: "pending",
  running: "running",
  paused: "paused",
  completed: "completed",
  failed: "failed",
} as const;

export interface PipelineWithNodes {
  id: number;
  name: string;
  yamlContent: string;
  status: PipelineWithNodesStatus;
  nodeCount: number;
  resolvedCount: number;
  nodes: PipelineNode[];
  createdAt: string;
  updatedAt: string;
}

export type ConsultationArguments = { [key: string]: unknown };

export type ConsultationStatus =
  (typeof ConsultationStatus)[keyof typeof ConsultationStatus];

export const ConsultationStatus = {
  PENDING: "PENDING",
  APPROVED: "APPROVED",
  REJECTED: "REJECTED",
} as const;

export interface Consultation {
  id: number;
  pipelineId: number;
  nodeId: string;
  approvalId: string;
  functionName: string;
  arguments: ConsultationArguments;
  message: string;
  status: ConsultationStatus;
  /** @nullable */
  feedback?: string | null;
  createdAt: string;
}

export type TelemetryEventPayload = { [key: string]: unknown };

export interface TelemetryEvent {
  id: number;
  /** @nullable */
  pipelineId?: number | null;
  /** @nullable */
  nodeId?: string | null;
  eventType: string;
  payload: TelemetryEventPayload;
  createdAt: string;
}

export interface CreatePipelineBody {
  name: string;
  yamlContent: string;
}

export type ResumePipelineBodyDecision =
  (typeof ResumePipelineBodyDecision)[keyof typeof ResumePipelineBodyDecision];

export const ResumePipelineBodyDecision = {
  approve: "approve",
  reject: "reject",
} as const;

export interface ResumePipelineBody {
  nodeId: string;
  decision: ResumePipelineBodyDecision;
  /** @nullable */
  feedback?: string | null;
}

export type ConsultationResponseBodyDecision =
  (typeof ConsultationResponseBodyDecision)[keyof typeof ConsultationResponseBodyDecision];

export const ConsultationResponseBodyDecision = {
  approve: "approve",
  reject: "reject",
} as const;

export interface ConsultationResponseBody {
  decision: ConsultationResponseBodyDecision;
  /** @nullable */
  feedback?: string | null;
}

export interface ErrorResponse {
  error: string;
}

export type ListTelemetryEventsParams = {
  pipelineId?: number;
  limit?: number;
};