# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Romeo PHD 6.0 - Agent Development Platform API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all pipelines
 */
export const ListPipelinesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  yamlContent: zod.string(),
  status: zod.enum(["pending", "running", "paused", "completed", "failed"]),
  nodeCount: zod.number(),
  resolvedCount: zod.number(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});
export const ListPipelinesResponse = zod.array(ListPipelinesResponseItem);

/**
 * @summary Create and start a new pipeline
 */
export const CreatePipelineBody = zod.object({
  name: zod.string(),
  yamlContent: zod.string(),
});

/**
 * @summary Get pipeline with all nodes and statuses
 */
export const GetPipelineParams = zod.object({
  id: zod.coerce.number(),
});

export const GetPipelineResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  yamlContent: zod.string(),
  status: zod.enum(["pending", "running", "paused", "completed", "failed"]),
  nodeCount: zod.number(),
  resolvedCount: zod.number(),
  nodes: zod.array(
    zod.object({
      id: zod.number(),
      pipelineId: zod.number(),
      nodeId: zod.string(),
      name: zod.string(),
      type: zod.string(),
      prompt: zod.string().nullish(),
      dependencies: zod.array(zod.string()),
      status: zod.enum([
        "PENDING",
        "RESOLVED",
        "NEEDS_CLARIFICATION",
        "AMBIGUOUS",
      ]),
      output: zod.string().nullish(),
      confidenceScore: zod.number().nullish(),
      positionX: zod.number(),
      positionY: zod.number(),
      executedAt: zod.date().nullish(),
    }),
  ),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Delete a pipeline
 */
export const DeletePipelineParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Resume pipeline after HITL review
 */
export const ResumePipelineParams = zod.object({
  id: zod.coerce.number(),
});

export const ResumePipelineBody = zod.object({
  nodeId: zod.string(),
  decision: zod.enum(["approve", "reject"]),
  feedback: zod.string().nullish(),
});

export const ResumePipelineResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  yamlContent: zod.string(),
  status: zod.enum(["pending", "running", "paused", "completed", "failed"]),
  nodeCount: zod.number(),
  resolvedCount: zod.number(),
  createdAt: zod.date(),
  updatedAt: zod.date(),
});

/**
 * @summary Start executing the pipeline nodes via SSE
 */
export const ExecutePipelineParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get all nodes for a pipeline
 */
export const ListPipelineNodesParams = zod.object({
  id: zod.coerce.number(),
});

export const ListPipelineNodesResponseItem = zod.object({
  id: zod.number(),
  pipelineId: zod.number(),
  nodeId: zod.string(),
  name: zod.string(),
  type: zod.string(),
  prompt: zod.string().nullish(),
  dependencies: zod.array(zod.string()),
  status: zod.enum(["PENDING", "RESOLVED", "NEEDS_CLARIFICATION", "AMBIGUOUS"]),
  output: zod.string().nullish(),
  confidenceScore: zod.number().nullish(),
  positionX: zod.number(),
  positionY: zod.number(),
  executedAt: zod.date().nullish(),
});
export const ListPipelineNodesResponse = zod.array(
  ListPipelineNodesResponseItem,
);

/**
 * @summary Get pending consultations for a pipeline
 */
export const ListConsultationsParams = zod.object({
  id: zod.coerce.number(),
});

export const ListConsultationsResponseItem = zod.object({
  id: zod.number(),
  pipelineId: zod.number(),
  nodeId: zod.string(),
  approvalId: zod.string(),
  functionName: zod.string(),
  arguments: zod.object({}).passthrough(),
  message: zod.string(),
  status: zod.enum(["PENDING", "APPROVED", "REJECTED"]),
  feedback: zod.string().nullish(),
  createdAt: zod.date(),
});
export const ListConsultationsResponse = zod.array(
  ListConsultationsResponseItem,
);

/**
 * @summary Respond to a HITL consultation request
 */
export const RespondToConsultationParams = zod.object({
  approvalId: zod.coerce.string(),
});

export const RespondToConsultationBody = zod.object({
  decision: zod.enum(["approve", "reject"]),
  feedback: zod.string().nullish(),
});

export const RespondToConsultationResponse = zod.object({
  id: zod.number(),
  pipelineId: zod.number(),
  nodeId: zod.string(),
  approvalId: zod.string(),
  functionName: zod.string(),
  arguments: zod.object({}).passthrough(),
  message: zod.string(),
  status: zod.enum(["PENDING", "APPROVED", "REJECTED"]),
  feedback: zod.string().nullish(),
  createdAt: zod.date(),
});

/**
 * @summary List recent telemetry events
 */
export const ListTelemetryEventsQueryParams = zod.object({
  pipelineId: zod.coerce.number().optional(),
  limit: zod.coerce.number().optional(),
});

export const ListTelemetryEventsResponseItem = zod.object({
  id: zod.number(),
  pipelineId: zod.number().nullish(),
  nodeId: zod.string().nullish(),
  eventType: zod.string(),
  payload: zod.object({}).passthrough(),
  createdAt: zod.date(),
});
export const ListTelemetryEventsResponse = zod.array(
  ListTelemetryEventsResponseItem,
);