# lib/api-zod/src/generated/types/consultation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Romeo PHD 6.0 - Agent Development Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ConsultationArguments } from "./consultationArguments";
import type { ConsultationStatus } from "./consultationStatus";

export interface Consultation {
  id: number;
  pipelineId: number;
  nodeId: string;
  approvalId: string;
  functionName: string;
  arguments: ConsultationArguments;
  message: string;
  status: ConsultationStatus;
  /** @nullable */
  feedback?: string | null;
  createdAt: Date;
}