# lib/api-zod/src/generated/types/consultationResponseBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Romeo PHD 6.0 - Agent Development Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ConsultationResponseBodyDecision } from "./consultationResponseBodyDecision";

export interface ConsultationResponseBody {
  decision: ConsultationResponseBodyDecision;
  /** @nullable */
  feedback?: string | null;
}