# lib/api-zod/src/generated/types/createPipelineBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Romeo PHD 6.0 - Agent Development Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreatePipelineBody {
  name: string;
  yamlContent: string;
}