# lib/api-zod/src/generated/types/pipelineWithNodes.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Romeo PHD 6.0 - Agent Development Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { PipelineNode } from "./pipelineNode";
import type { PipelineWithNodesStatus } from "./pipelineWithNodesStatus";

export interface PipelineWithNodes {
  id: number;
  name: string;
  yamlContent: string;
  status: PipelineWithNodesStatus;
  nodeCount: number;
  resolvedCount: number;
  nodes: PipelineNode[];
  createdAt: Date;
  updatedAt: Date;
}